added file logging to Contract and Zeebe project
