+++ amorphie.contract.zeebe/Modules/ZeebeDocumentUpload.cs
@@ -1,25 +1,13 @@
-using System.Security.Cryptography.X509Certificates;
-using System.Xml.XPath;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Reflection.Metadata;
-using System.Text;
-using System.Threading.Tasks;
 using amorphie.contract.core.Entity.Document;
-using amorphie.contract.zeebe.Service.Minio;
 using amorphie.contract.data.Contexts;
 using amorphie.contract.zeebe.Model;
 using Dapr.Client;
-using Google.Api;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 using Microsoft.OpenApi.Models;
-using System.ComponentModel;
-using Google.Protobuf;
 using amorphie.contract.zeebe.Services;
 using Newtonsoft.Json;
 using amorphie.contract.core.Enum;
+using amorphie.contract.core.Services;
 
 namespace amorphie.contract.zeebe.Modules
 {
