+++ amorphie.contract.zeebe/Program.cs
@@ -2,10 +2,13 @@ using amorphie.contract.data.Contexts;
 using amorphie.contract.core;
 using amorphie.contract.zeebe.Modules;
 using amorphie.contract.zeebe.Modules.ZeebeDocumentDef;
-using amorphie.contract.zeebe.Service.Minio;
 using amorphie.contract.zeebe.Services;
 using Microsoft.EntityFrameworkCore;
 using amorphie.contract.application;
+using amorphie.core.Extension;
+using Elastic.Apm.NetCoreAll;
+using amorphie.contract.core.Services;
+using amorphie.contract.data.Services;
 
 var builder = WebApplication.CreateBuilder(args);
 IConfiguration Configuration;
@@ -21,6 +24,8 @@ Configuration = builder
 // Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
 builder.Services.AddEndpointsApiExplorer();
 builder.Services.AddSwaggerGen();
+builder.AddSeriLog();
+
 builder.Services.AddDbContext<ProjectDbContext>
     (options => options.UseNpgsql(Configuration.GetConnectionString("DefaultConnection")));
 builder.Services.AddCors(options =>
@@ -34,6 +39,7 @@ builder.Services.AddCors(options =>
         });
 });
 builder.Services.AddDaprClient();
+
 builder.Services.AddSingleton<IConfigurationRoot>(provider => builder.Configuration);
 builder.Services.Configure<AppSettings>(builder.Configuration.GetSection("AppSettings"));
 var settings = builder.Configuration.Get<AppSettings>();
@@ -47,6 +53,7 @@ builder.Services.AddScoped<IContractDefinitionService, ContractDefinitionService
 
 builder.Services.AddApplicationServices();
 
+
 var app = builder.Build();
 using var scope = app.Services.CreateScope();
 var db = scope.ServiceProvider.GetRequiredService<ProjectDbContext>();
@@ -58,32 +65,11 @@ app.UseSwaggerUI();
 
 app.UseHttpsRedirection();
 
-var summaries = new[]
-{
-    "Freezing", "Bracing", "Chilly", "Cool", "Mild", "Warm", "Balmy", "Hot", "Sweltering", "Scorching"
-};
+app.UseAllElasticApm(builder.Configuration);
+
 app.MapZeebeDocumentUploadEndpoints();
 app.MapZeebeDocumentDefinitionEndpoints();
 app.MapZeebeContractDefinitionEndpoints();
 app.MapZeebeDocumentGroupDefinitionEndpoints();
-app.MapGet("/weatherforecast", () =>
-{
-    var forecast = Enumerable.Range(1, 5).Select(index =>
-       new WeatherForecast
-       (
-           DateOnly.FromDateTime(DateTime.Now.AddDays(index)),
-           Random.Shared.Next(-20, 55),
-           summaries[Random.Shared.Next(summaries.Length)]
-       ))
-        .ToArray();
-    return forecast;
-})
-.WithName("GetWeatherForecast")
-.WithOpenApi();
 app.Run();
 
-
-record WeatherForecast(DateOnly Date, int TemperatureC, string? Summary)
-{
-    public int TemperatureF => 32 + (int)(TemperatureC / 0.5556);
-}
