+++ amorphie.contract.zeebe/amorphie.contract.zeebe.csproj
@@ -19,13 +19,14 @@
     <PackageReference Include="Swashbuckle.AspNetCore" Version="6.5.0" />
     <PackageReference Include="Microsoft.EntityFrameworkCore" Version="7.0.13" />
     <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="7.0.11" />
-    <PackageReference Include="amorphie.core" Version="1.1.4" />
-    <PackageReference Include="amorphie.core.security" Version="1.0.0" />
     <PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="7.0.11" />
     <PackageReference Include="StackExchange.Redis" Version="2.6.122" />
     <ProjectReference Include="..\amorphie.contract.data\amorphie.contract.data.csproj" />
     <ProjectReference Include="..\amorphie.contract.core\amorphie.contract.core.csproj" />
     <ProjectReference Include="..\amorphie.contract.application\amorphie.contract.application.csproj" />
+    <PackageReference Include="amorphie.core" Version="1.1.25" />
+    <PackageReference Include="amorphie.core.security" Version="1.0.0" />
+    <PackageReference Include="Elastic.Apm.NetCoreAll" Version="1.16.1" />
 
   </ItemGroup>
