+++ amorphie.contract.zeebe/appsettings.Test.json
@@ -1,10 +1,4 @@
 {
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Warning"
-    }
-  },
   "ConnectionStrings": {
     "DefaultConnection": "DB_CONNECTION"
   },
@@ -13,12 +7,63 @@
     "HtmlRenderEndpoint": "TemplateEngineHtmlRenderEndpointValue",
     "PdfRenderEndpoint": "TemplateEnginePdfRenderEndpointValue",
     "GetTemplateEndpoint": "TemplateEngineGetTemplateEndpointValue"
-  }
-  ,
+  },
   "minio": {
     "EndPoint": "EndPointValue",
     "AccessKey": "AccessKeyValue",
     "SecretKey": "SecretKeyValue",
     "BucketName": "BucketNameValue"
+  },
+  "ElasticApm": {
+    "SecretToken": "ElasticApm_SecretToken",
+    "ServerUrl": "ElasticApm_URL",
+    "ServiceName": "amoprhie-contract-zeebe",
+    "Environment": "Test",
+    "TransactionSampleRate": 1.0,
+    "TransactionIgnoreUlrs": "/health"
+  },
+  "MetricsOptions": {
+    "DefaultContextLabel": "amoprhie-contract-zeebe",
+    "Enabled": true
+  },
+  "MetricsWebTrackingOptions": {
+    "ApdexTrackingEnabled": true,
+    "ApdexTSeconds": 0.1,
+    "IgnoredHttpStatusCodes": [],
+    "IgnoredRoutesRegexPatterns": [],
+    "OAuth2TrackingEnabled": true
+  },
+  "MetricEndpointsOptions": {
+    "MetricsEndpointEnabled": true,
+    "MetricsTextEndpointEnabled": true,
+    "EnvironmentInfoEndpointEnabled": true
+  },
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft.AspNetCore": "Warning"
+    }
+  },
+  "Serilog": {
+    "MinimumLevel": {
+      "Default": "Information",
+      "Override": {
+        "Microsoft": "Warning",
+        "Microsoft.AspNetCore": "Warning",
+        "Microsoft.AspNetCore.Hosting.Diagnostics": "Warning",
+        "Microsoft.AspNetCore.HttpLogging.HttpLoggingMiddleware": "Information",
+        "System": "Warning"
+      }
+    },
+    "WriteTo": [
+      {
+        "Name": "File",
+        "Args": {
+          "path": "./logs/log-amorphie-contract-zeebe-.json",
+          "formatter": "Serilog.Formatting.Compact.CompactJsonFormatter, Serilog.Formatting.Compact",
+          "rollingInterval": "Day"
+        }
+      }
+    ]
   }
 }
