+++ amorphie.contract/appsettings.Prod.json
@@ -14,50 +14,13 @@
     "SecretKey": "SecretKeyValue",
     "BucketName": "BucketNameValue"
   },
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Warning"
-    }
-  },
-  "AllowedHosts": "*",
-  "Serilog": {
-    "Using": [
-      "Serilog.Sinks.Console",
-      "Serilog.Sinks.ElasticSearch"
-    ],
-    "MinimumLevel": {
-      "Default": "Information",
-      "Override": {
-        "Microsoft": "Warning",
-        "System": "Warning"
-      }
-    },
-    "WriteTo": [
-      {
-        "Name": "Console"
-      },
-      {
-        "Name": "Debug"
-      },
-      {
-        "Name": "Elasticsearch",
-        "Args": {
-          "nodeUris": "ElasticSearch_URL",
-          "indexFormat": "amoprhie-contract-index-{0:yyyy.MM}",
-          "emitEventFailure": "WriteToSelfLog",
-          "autoRegisterTemplate": true,
-          "connectionGlobalHeaders": "Authorization=Basic ElasticSearch_AuthToken"
-        }
-      }
-    ]
-  },
   "ElasticApm": {
     "SecretToken": "ElasticApm_SecretToken",
     "ServerUrl": "ElasticApm_URL",
-    "ServiceName": "LagerthaApi",
-    "Environment": "Test",
-    "TransactionSampleRate": 1.0
+    "ServiceName": "amoprhie-contract",
+    "Environment": "Production",
+    "TransactionSampleRate": 1.0,
+    "TransactionIgnoreUlrs": "/health"
   },
   "MetricsOptions": {
     "DefaultContextLabel": "amoprhie-contract",
@@ -74,5 +37,33 @@
     "MetricsEndpointEnabled": true,
     "MetricsTextEndpointEnabled": true,
     "EnvironmentInfoEndpointEnabled": true
+  },
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft.AspNetCore": "Warning"
+    }
+  },
+  "Serilog": {
+    "MinimumLevel": {
+      "Default": "Information",
+      "Override": {
+        "Microsoft": "Warning",
+        "Microsoft.AspNetCore": "Warning",
+        "Microsoft.AspNetCore.Hosting.Diagnostics": "Warning",
+        "Microsoft.AspNetCore.HttpLogging.HttpLoggingMiddleware" : "Information",
+        "System": "Warning"
+      }
+    },
+    "WriteTo": [
+      { 
+        "Name": "File", 
+        "Args": { 
+          "path": "./logs/log-amorphie-contract-.json", 
+          "formatter": "Serilog.Formatting.Compact.CompactJsonFormatter, Serilog.Formatting.Compact",
+          "rollingInterval": "Day" 
+        } 
+      }
+    ]
   }
 }
